==> code/00-converting-json-to-csv.ipynb <==
# Convert JSON to CSV

The file extracted from the Yelp dataset (`yelp_academic.tar`) is a typeless file (`yelp_dataset`). This is actually another tar file, so manually add the `.tar` extension and extract again to get to the true data files. 

The Yelp Dataset is a series of JSON files. I will convert the JSON files to CSVs.

### Import Libaries

import pandas as pd
import time
import csv
import simplejson as sjson
import collections

### JSON and CSV File Paths

categories = ['business', 'checkin', 'photo', 'review', 'tip', 'user']

json_file_paths = {cat : f'../data/external/yelp_academic_dataset_{cat}.json'
                   for cat in categories}

csv_file_paths = {cat : f'../data/raw/yelp_academic_dataset_{cat}.csv'
                   for cat in categories}

json_file_paths

csv_file_paths

### Read in JSON, Write out to CSV

def read_and_write_file(json_file_path, csv_file_path, column_names):
    '''
    Arguments:
    json_file_path (str) : pathname of a JSON file
    csv_file_path (str) : pathname of CSV file
    column_names (set) : column names (keys) of a JSON file
    
    Returns:
    None
    
    Read in JSON file
    Write out as CSV file with column_names as header
    '''
    
    with open(csv_file_path, 'w+') as f_out:
        csv_file = csv.writer(f_out)
        csv_file.writerow(list(column_names))
        with open(json_file_path) as f_in:
            for line in f_in:
                d = sjson.loads(line)
                csv_file.writerow(get_row(d, column_names))
                
                


### Get Column Names

def get_column_names(d, parent_key = ''):
    '''
    Arguments:
    d (dict) : Nested dictionary of contents

    Recursively extracts the keys of d and 
    returns a list of flattened keys as column_names

    Returns:
    column_names (list) : Flattened list of key names
    
    Example: 
    d = {'a' : {'b' : 0, 'c' : 1}}
    column_names = ['a.b', 'a.c']
    '''
    
    column_names = []
    
    for key, val in d.items():
        column_name = f'{parent_key}.{key}' if parent_key else key
        
        # if val is a dictionary, recursively call get_column_names
        if isinstance(val, collections.MutableMapping):
            column_names.extend(get_column_names(val, column_name))
        else:
            column_names.append(column_name)
            
    return column_names

d = {'a' : {'b' : 0, 'c' : 1}}
get_column_names(d)

### Get Nested Value

def get_nested_value(d, key):
    '''
    Arguments: 
    d (dict) : Nested dictionary of contents
    key (str) : Flattened key string
    
    Recursively extracts the value from a nested dictionary given a key
    
    Returns:
    Value of d[key]
    
    Example:
    d = {'a' : {'b' : 0, 'c' : 1}}
    key = 'a.b'
    value = 0
    '''
    
    if d is None:
        return None
    
    # check if key is an end node
    if '.' not in key:
        # return value = d[key] if it exists
        if key not in d:
            return None
        return d[key]
    
    # extract parent key from child key(s)
    parent_key, child_key = key.split('.', 1)
    
    if parent_key not in d:
        return None
    # get value of d[parent_key] is a dictionary for recursive call
    child_d = d[parent_key]
    
    return get_nested_value(child_d, child_key)


d = {'a' : {'b' : 0, 'c' : 1}, 'd' : {'e' : None, 'f' : 2}}
get_nested_value(d, 'd.f')

### Get Row

def get_row(d, column_names):
    '''
    Arguments:
    d : Nested dictionary of contents
    column_names : List of column names. ** ASSUMES ALL column_names ARE VALID **
     
    Returns:
    row (list) : Values as strings
    
    Returns a csv compatible row given column names and a dict
    '''
    row = []    
    for column_name in column_names:
        val = get_nested_value(d, column_name)

        if val is not None:
            row.append(str(val))
        else:
            row.append('')

    return row
    

d = {'a' : {'b' : 0, 'c' : 'C'}}
get_row(d, ['a', 'b', 'a.b', 'a.c', 'a.d'])

### Get Superset of Column Names
The superset is the set of all possible column names.

def get_superset_of_column_names(json_file_path):
    '''
    Arguments:
    json_file_path (str) : pathname of a JSON file
    
    Extracts the unique column names (keys) of a JSON file
    
    Returns:
    column_names (set) : Column names (keys) of a JSON file
    
    Example:
    json_file = {'a' : {'b' : 1, 'c' : 2}}
    column_names = {'a.b', 'a.c'}
    '''
    
    column_names = set()
    with open(json_file_path) as f_in:
        for line in f_in:
            d = sjson.loads(line)
            column_names.update(set(get_column_names(d)))

    return column_names

#### Business column names

%%time

json_file_path = '../data/external/yelp_academic_dataset_business.json'
column_names = get_superset_of_column_names(json_file_path)
print(column_names)

### Review column names

%%time 

json_file_path = '../data/external/yelp_academic_dataset_review.json'
column_names = get_superset_of_column_names(json_file_path)
print(column_names)

## Convert JSON to CSV

#### Business

%%time

json_file_path = json_file_paths['business']
csv_file_path = csv_file_paths['business']

column_names = get_superset_of_column_names(json_file_path)
print(column_names)

read_and_write_file(json_file_path, csv_file_path, column_names)

#### Checkin

%%time

json_file_path = json_file_paths['checkin']
csv_file_path = csv_file_paths['checkin']

column_names = get_superset_of_column_names(json_file_path)
print(column_names)

read_and_write_file(json_file_path, csv_file_path, column_names)

#### Photo

t0 = time.time()

json_file_path = json_file_paths['photo']
csv_file_path = csv_file_paths['photo']

column_names = get_superset_of_column_names(json_file_path)
print(column_names)

read_and_write_file(json_file_path, csv_file_path, column_names)

print(time.time() - t0)

#### Review

t0 = time.time()

json_file_path = '../data/external/yelp_academic_dataset_review.json'
csv_file_path = json_file_path.replace('json', 'csv')

column_names = get_superset_of_column_names(json_file_path)
read_and_write_file(json_file_path, csv_file_path, column_names)

print(time.time() - t0)

#### Tip

t0 = time.time()

json_file_path = json_file_paths['tip']
csv_file_path = csv_file_paths['tip']

column_names = get_superset_of_column_names(json_file_path)
print(column_names)

read_and_write_file(json_file_path, csv_file_path, column_names)

print(time.time() - t0)

#### User

t0 = time.time()

json_file_path = json_file_paths['user']
csv_file_path = csv_file_paths['user']

column_names = get_superset_of_column_names(json_file_path)
print(column_names)

read_and_write_file(json_file_path, csv_file_path, column_names)

print(time.time() - t0)

### Load CSV

Verify that JSON file was successfully converted to CSV.

user_df = pd.read_csv(csv_file_paths['user'])

user_df.head()

user_df.shape
