==> code/01-cleaning-data.ipynb <==
# Cleaning Data

## Table of Contents

1. [Find restaurants in Business data](#Find-restaurants-in-Business-data)
2. [Select restaurant reviews from Review data](#Select-restaurant-reviews-from-Review-data)
3. [Clean Review Data](#Clean-Review-Data)
4. [Clean Restaurant data](#Clean-restaurant-data)



### Import Libraries

import csv
import os
import pandas as pd
from time import time

### Set filepaths

raw_data_directory     = os.path.join('..', 'data', 'raw')
interim_data_directory = os.path.join('..', 'data', 'interim')

review_filepath            = os.path.join(raw_data_directory, 'yelp_academic_dataset_review.csv')
business_filepath          = os.path.join(raw_data_directory, 'yelp_academic_dataset_business.csv')
restaurant_review_filepath = os.path.join(interim_data_directory, 'restaurant_review.csv')
restaurant_filepath        = os.path.join(interim_data_directory, 'restaurant.csv')

### Find restaurants in Business data
Restaurants are businesses with "Restaurants" in the `categories` column.

%%time

restaurant_ids = set()

with open(restaurant_filepath, mode = 'w', encoding = 'utf_8') as f_out:
    with open(business_filepath, encoding = 'utf_8') as f_in:

        reader = csv.DictReader(f_in)
        columns = reader.fieldnames
        
        writer = csv.DictWriter(f_out, fieldnames = columns)
        writer.writeheader()

        # iterate through each line in the file
        for row in reader:

            # skip businesses that are not restaurants
            try: 
                if 'Restaurants' not in row['categories']:
                    continue
            except:
                continue

            # add business_id to restaurant_ids set
            restaurant_ids.add(row['business_id'])
            
            # write row
            writer.writerow(row)            

restaurant_ids = frozenset(restaurant_ids)

print (f'Found {len(restaurant_ids)} restaurants.')

### Select restaurant reviews from Review data

Save restaurant reviews to a separate file.

%%time

# Change to True to run this code block. Estimated runtime: 3 minutes
run = False

if run:
    num_reviews = 0

    # Write restaurant reviews to csv file
    with open(restaurant_review_filepath, 'w', encoding = 'utf_8') as f_out:

        # Open all reviews csv file
        with open(review_filepath, encoding = 'utf_8') as f_in:
            # Instantiate reader
            reader = csv.DictReader(f_in)

            # Get column names
            columns = reader.fieldnames

            # Instantiate writer
            writer = csv.DictWriter(f_out, fieldnames = columns)

            # Write column names
            writer.writeheader()

            # Loop through all reviews
            for row in reader:            
                # Skip reviews that are not about a restaurant
                if row['business_id'] not in restaurant_ids:
                    continue

                # Write row    
                writer.writerow(row)
                num_reviews += 1

    print (f'Found {num_reviews} restaurant reviews.')


#### Load Restaurant Review data

review_df = pd.read_csv(restaurant_review_filepath)
review_df.head()

### Get the reviewed restaurants

business_df = pd.read_csv(business_filepath)

business_df.nlargest(5, 'review_count')[['business_id', 'name', 'review_count']]

#### Cross reference with review data

review_df['business_id'].value_counts().head()

For the rest of this project, I will use Mon Ami Gabi.

### Clean Review Data

#### Check for missing data

review_df.isna().sum()

There is one review with missing text, so let's remove it.

#### Drop rows with missing reviews

review_df[review_df['text'].isna()]

review_df.drop(review_df[review_df['text'].isna()].index, axis = 0, inplace = True)

review_df[review_df['text'].isna()]

#### Add actual `business_name` to Review Data

# Build a dictionary to map business_id : business_name 
business_id_name_mapper = dict(zip(business_df['business_id'], business_df['name']))

%%time
# add business name to dataframe
review_df['business_name'] = review_df['business_id'].map(business_id_name_mapper)

review_df.head()

#### Remove `\n` characters

review_df['text'] = review_df['text'].apply(lambda row : row.replace('\n', ' '))

# Confirm '\n' characters are removed
review_df[review_df['text'].str.contains('\n')]

#### Keep relevant columns

review_df = review_df[['date', 'stars', 'text', 'review_id', 'business_id', 'business_name']]

review_df[review_df['business_name']=='Euro Shawarma']

review_df.head()

review_df.shape

#### Save clean review data

%%time
review_df.to_csv(restaurant_review_filepath, index = False)

### Clean restaurant data

#### Load restaurant data

restaurant_df = pd.read_csv(restaurant_filepath)

restaurant_df.head()

restaurant_df.shape

restaurant_df.columns

### Cleaning Restaurant Data

#### Drop rows with missing `postal_code` or `city`

restaurant_df['postal_code'].isna().sum() / restaurant_df.shape[0]

restaurant_df['city'].isna().sum() / restaurant_df.shape[0]

0.2% of restaurants are missing postal code or city data. I will remove them.

restaurant_df = restaurant_df.dropna(subset=['postal_code', 'city'])

len(restaurant_df)

#### Keep relevant columns

columns = ['name', 'business_id', 'stars', 'review_count', 
           'categories', 'longitude', 'latitude', 'postal_code', 'city', 'state']

restaurant_df = restaurant_df[columns]


restaurant_df.head()

#### Last check for missing values

restaurant_df.isna().sum()

#### Save clean restaurant data

restaurant_df.to_csv(restaurant_filepath, index = False)
