==> code/03-selecting-a-restaurant.ipynb <==
# Subsetting Data

### Import Libraries

import csv
import os
import pandas as pd
from time import time
import warnings

# For reading from Postgres
from sqlalchemy import create_engine

warnings.filterwarnings('ignore')

### Set filepaths

raw_data_directory     = os.path.join('..', 'data', 'raw')
interim_data_directory = os.path.join('..', 'data', 'interim')

review_filepath            = os.path.join(raw_data_directory, 'yelp_academic_dataset_review.csv')
business_filepath          = os.path.join(raw_data_directory, 'yelp_academic_dataset_business.csv')
restaurant_review_filepath = os.path.join(interim_data_directory, 'restaurant_review.csv')
restaurant_filepath        = os.path.join(interim_data_directory, 'restaurant.csv')

### Load Data

%%time 
review_df = pd.read_csv(restaurant_review_filepath)

review_df.head()

restaurant_df = pd.read_csv(restaurant_filepath)

restaurant_df.head()

mon_ami_gabi_id = '4JNXUYY8wbaaDmk3BPzlWw'

### Selecting a restaurant using PostgreSQL

public_ip = 'localhost'
username = 'postgres'
password = 'password'
port = '5432'
database = 'yelp'

# Construct database URL from environment variables
uri = f'postgresql://{username}:{password}@{public_ip}:{port}/{database}'

# Connection to Postgres database
engine = create_engine(uri)

%%time

SQL = f'''
SELECT date, stars, text, review_id, business_name
FROM reviews
WHERE business_id = '{mon_ami_gabi_id}'
'''

reviews = pd.read_sql(SQL, con = engine)

reviews.head()

reviews.shape

### Building my own search engine 

class SearchCriteria:
    '''
    query (any): value to search for
    column (str): column name of dataframe to search
    exact (bool): True: return exact matches
                  False: return results that contains the query
    '''
    
    query = None
    column = None
    exact = None
    
    def __init__(self, query, column, exact = False):
        
        assert type(exact) is bool, 'TypeError: exact must be boolean.'
        assert type(column) is str, 'TypeError: column must be string.'
        
        self.query = query
        self.column = column
        self.exact = exact
        
        
    def summarize(self):
        
        print(f'Query: {self.query}\nColumn: {self.column}\nExact: {self.exact}\n')
        
    

def searchCriteria(query, column, exact = False):
    '''
    Returns a dictionary for a single query criteria.
    '''
    if type(exact) is not bool:
        print('Error: Exact must be True or False.')
        return None
    
    return {
        'query' : query,
        'column' : column,
        'exact' : exact
    }

### TODO: Optimize search_reviews function
Function takes 63 seconds vs   
Iterative Masking takes 0.48 seconds

def search_reviews(df, queries):
    '''
    Returns a subset of df that matches a list of SearchCriteria.
    
    df: Dataframe
    queries: 
    
    Use the SearchCriteria class to build a list of queries.
    '''
    supermask = True

    for query in queries:
        if query.exact:
            mask = df[query.column] == query.query
        else:
            mask = df[query.column].str.contains(query.query, case = False)
        supermask = supermask & mask

    return df[supermask]                


### Select Mon Ami Gabi reviews

mon_ami_gabi_reviews = review_df[review_df['business_id'] == mon_ami_gabi_id]

mon_ami_gabi_reviews.shape

[Mon Ami Gabi's Yelp Page](https://www.yelp.com/biz/mon-ami-gabi-las-vegas-2)

mon_ami_gabi_reviews.to_csv('../data/interim/mon_ami_gabi_reviews.csv', index = False)

mon_ami_gabi_reviews.head()

### Select reviews from Mon Ami Gabi that mention a certain menu item

#### Select Mon Ami Gabi reviews that mention French `onion soup`

I tried subsetting using three methods: 
1. A search function for reusable code
2. Iterative masking: `df1 = df[mask1]`, `df2 = df1[mask2]`, etc.
3. Consecutive masking: `df[mask1 & mask2 & ...]`

Iterative masking performed the fastest.

queries = []
queries.append(SearchCriteria(mon_ami_gabi_id, 'business_id'))
queries.append(SearchCriteria('onion soup', 'text'))

# Using search_reviews function
t0 = time()

onion_soup_reviews = search_reviews(mon_ami_gabi_reviews, queries)

t1 = np.round (time()-t0, 4)

print(f'Found {onion_soup_reviews.shape[0]} results in {t1} seconds')

%%time
mon_ami_gabi_mask = review_df['business_id'] == mon_ami_gabi_id
onion_soup_mask = mon_ami_gabi_reviews['text'].str.contains('onion soup', case = False)


# Searching Mon Ami Gabi AND Onion Soup
t0 = time()

onion_soup_reviews = review_df[mon_ami_gabi_mask & onion_soup_mask]

t1 = np.round(time() - t0, 4)

print(f'Found {onion_soup_reviews.shape[0]} results in {t1} seconds')

# Searching Onion Soup in Mon Ami Gabi data
t0 = time()

onion_soup_reviews = mon_ami_gabi_reviews[onion_soup_mask]

t1 = np.round(time() - t0, 4)

print(f'Found {onion_soup_reviews.shape[0]} results in {t1} seconds')

# Searching Mon Ami Gabi, then searching Onion Soup
t0 = time()

mon_ami_gabi = review_df[mon_ami_gabi_mask]
onion_soup_reviews = mon_ami_gabi[onion_soup_mask]

t1 = np.round(time() - t0, 4)

print(f'Found {onion_soup_reviews.shape[0]} results in {t1} seconds')

# chain masking
# Searching Mon Ami Gabi, then searching Onion Soup
t0 = time()

onion_soup_reviews = review_df[mon_ami_gabi_mask][onion_soup_mask]

t1 = np.round(time() - t0, 4)

print(f'Found {onion_soup_reviews.shape[0]} results in {t1} seconds')

Chain masking seems to be the fastest method. However it requires the entire dataset to be loaded into memory, which is not scalable. 

onion_soup_reviews.shape

onion_soup_reviews.head()

#### Save Onion Soup reviews

onion_soup_reviews.to_csv('../data/interim/onion_soup_reviews.csv', index = False)

#### Select Mon Ami Gabi reviews that  mention eggs `benedict`

t0 = time()

eggs_benedict_reviews = mon_ami_gabi_reviews[mon_ami_gabi_reviews['text'].str.contains('benedict', case = False)]
t1 = np.round(time() - t0, 2)

print(f'Found {eggs_benedict_reviews.shape[0]} results in {t1} seconds')

#### Save Eggs Benedict reviews

eggs_benedict_reviews.to_csv('../data/interim/eggs_benedict_reviews.csv', index = False)

### Select restaurants with the most reviews

popular_restaurants = restaurant_df.nlargest(1000, 'review_count')
popular_restaurants.head(10)

Vegas loves their restaurants.

#### Save 1000 most reviewed restaurants

popular_restaurants.to_csv('../data/interim/popular_restaurants.csv')

### Select restaurants with the lowest stars

worst_restaurants = popular_restaurants.nsmallest(1000, 'stars')

worst_restaurants.head(10)

### Trade offs

**Speed comparison:**  
PostgreSQL: `7.13 s`  
Loading dataset into memory: `32.8 s`
My custom search engine: `0.1076 s`  
Chain masking: `0.0111 s`  

Verdict: SQL is the way to go. Loading the entire dataset into memory is not scalable. The query takes marginally less time if you're only doing one query. If making multiple similar queries, use a database system that caches queries, such as Google BigQuery.
