==> code/04.2-splitting-to-sentences.ipynb <==
## Splitting Reviews to Sentences

The input of this notebook is a dataframe of reviews that mention one menu item. The output is a dataframe of the individual sentences that mention that menu item. As an example, I am using `onion soup` and `eggs benedict`.

### Import Libraries

import pandas as pd
import numpy as np
from time import time

# Display full content of column
# pd.set_option('display.max_colwidth', -1)

import spacy
import textblob

# For reading from Postgres
from sqlalchemy import create_engine

# For pickling
import pickle

# For tracking progress
from IPython.display import clear_output

# For reading and writing to postgres
from odo import odo

# For detecting language of document
from langdetect import detect
from langdetect import DetectorFactory 

# for consistent results
DetectorFactory.seed = 42 

import nltk

# The NLP workhorse in Python is Natural Language Toolkit (NLTK)
# Tokenizing, lemmatizing
from nltk import word_tokenize, pos_tag, ne_chunk

# Preprocessing packages used in class
from nltk.stem import WordNetLemmatizer
from nltk.tokenize import RegexpTokenizer


# For loading secret environment variables, e.g. postgres username and password
import os
from dotenv import load_dotenv, find_dotenv


### Create Postgres table

sql = '''
CREATE TABLE reviews (
    date          date, 
    stars         integer NOT NULL,
    text          varchar(5000), 
    review_id     varchar(22),
    business_id   varchar(22),
    business_name varchar(64)  
);
'''

### Import CSV file to Postgres

raw_data_directory     = os.path.join('..', 'data', 'raw')
interim_data_directory = os.path.join('..', 'data', 'interim')

review_filepath            = os.path.join(raw_data_directory, 'yelp_academic_dataset_review.csv')
business_filepath          = os.path.join(raw_data_directory, 'yelp_academic_dataset_business.csv')
restaurant_review_filepath = os.path.join(interim_data_directory, 'restaurant_review.csv')
restaurant_filepath        = os.path.join(interim_data_directory, 'restaurant.csv')

sql = f'''
COPY reviews(date, stars, text, review_id, business_id, business_name)
FROM '{restaurant_review_filepath}' DELIMITER ',' CSV HEADER;
'''

### Set Environment Variables

# Find .env
dotenv_path = find_dotenv()

# Load entries as environment variables
load_dotenv(dotenv_path)

public_ip = os.environ.get("PUBLIC_IP")
username = os.environ.get("USERNAME")
password = os.environ.get("PASSWORD")
port = os.environ.get("PORT")

# Construct database URL from environment variables
uri = f'postgresql://{username}:{password}@{public_ip}:{port}'

### Load Data 
#### Option 1: Local Postgres

public_ip = 'localhost'
username = 'postgres'
password = 'password'
port = '5432'
database = 'yelp'

# Construct database URL from environment variables
uri = f'postgresql://{username}:{password}@{public_ip}:{port}/{database}'

# uri = 'postgresql://pos:lilo8catfood@127.0.0.0:5432'
# reviews = odo(uri+'::reviews', pd.DataFrame)

# Connection to Postgres database
engine = create_engine(uri)


%%time

SQL = '''
SELECT r.*
FROM reviews AS r
WHERE r.business_name ILIKE 'Mon Ami Gabi'
  AND r.text LIKE '%%onion soup%%'
'''

onion_soup_reviews = pd.read_sql(SQL, con = engine)

onion_soup_reviews.shape

onion_soup_reviews.head()

#### Option 2: Load data locally

onion_soup_reviews = pd.read_csv('../data/interim/onion_soup_reviews.csv')
onion_soup_reviews.shape

eggs_benedict_reviews = pd.read_csv('../data/interim/eggs_benedict_reviews.csv')
eggs_benedict_reviews.shape

menu = pickle.load( open( "../data/interim/mon_ami_gabi_menu.pk", "rb" ) )

menu.head()

### Drop rows that are not English

def get_english_reviews(df):
    language = df['text'].apply(detect)
                  
    return df.drop(df[language != 'en'].index, axis = 0)
    

%%time
onion_soup_reviews = get_english_reviews(onion_soup_reviews)
print(onion_soup_reviews.shape[0])

%%time
eggs_benedict_reviews = get_english_reviews(eggs_benedict_reviews)
print(eggs_benedict_reviews.shape[0])

### Tokenize & Lemmatize Data

[Penn Part of Speech Tags](https://cs.nyu.edu/grishman/jet/guide/PennPOS.html)

def preprocess2(sentence):
    # Tokenize doc
    tokens = word_tokenize(sentence)
    
    # Tag sentences
    tagged_tokens = pos_tag(tokens)

    # Named entity chunker
    ne_chunked_tokens = ne_chunk(tagged_tokens, binary = True)
#     return ne_chunked_tokens
    # Extract all named entities
    named_entities = []
    
    for tagged_tree in ne_chunked_tokens:
#         print(tagged_tree)
        if hasattr(tagged_tree, 'label'):
            entity_name = ' '.join(c[0] for c in tagged_tree.leaves())
            entity_type = tagged_tree.label() # category
            named_entities.append((entity_name, entity_type))
    
    return named_entities
    

processed_review = preprocess2(onion_soup_reviews['text'][0])

processed_review

NLTK's tokenizer fails to detect onion soup as an entity, so I will continue using a less programmatic approach. 

### Separate each review document into individual sentences

def flatten(superlist): 
    '''
    Arguments: 
    superlist : A list of list of strings.

    Requirements: 
    Each element in superlist must be a list.
    
    Return:
    A flattened list of strings.

    ex: 
    flatten([['a'], ['b', 'c'], ['d', 'e', 'f']])
    >> ['a', 'b', 'c', 'd', 'e', 'f']
    '''    
    return [item \
            for sublist in superlist \
            for item in sublist]

def get_sentences(doc, menu_item):
    '''
    Arguments: 
    doc : pd.Series of reviews
    menu_item : pd.Series of a menu item. example: menu.iloc[0] = "onion_soup_au_gratin"
    
    Splits a string into individual sentences and
    selects only the sentences that contain the search term
    
    Return:
    DataFrame of sentences with their target
    '''
    sentences = doc.apply(lambda text : text.split('.'))
#     df['sentences'] = df['sentences'].apply(lambda sentences : [s for s in sentences])
    sentences = flatten(sentences)
    
    tagged_sentences = []
    sentences_tags = []
    
    n_sentences = len(sentences)
    
    for i,s in enumerate(sentences):
        s = s.lower()
        
        if (i+1)%1000==0:
            clear_output(wait = True)
            print(f'Finding tags in {i+1}/{n_sentences} sentences')

        tags = []
        for i,row in menu.iterrows():
            for variation in row['variations']:
                if variation in s:
                    # print(row['entity'], '\t', variation)
                    tags.append(row['name'])
                    break
        sentences_tags.append(', '.join(tags))
    
    return pd.DataFrame(list(zip(sentences, sentences_tags)), columns = ['text', 'tags'])
#     return pd.DataFrame([{'text' : s, 'target' : term} for s in sentences])    
    

### Find all sentences that mention "onion soup"

%%time
onion_soup_sentences = get_sentences(onion_soup_reviews['text'], menu[menu['id'] == 'onion_soup_au_gratin'])

onion_soup_sentences = onion_soup_sentences[onion_soup_sentences['tags'].str.contains('onion soup au gratin')]
onion_soup_sentences.shape


onion_soup_sentences.head()

onion_soup_sentences.to_csv('../data/interim/onion_soup_sentences.csv', index = False)


### Find all sentences that mention "eggs benedict" 

eggs_benedict_sentences = get_sentences(eggs_benedict_reviews['text'], menu[menu['id'] == 'classic_eggs_benedict'])

eggs_benedict_sentences = eggs_benedict_sentences[eggs_benedict_sentences['tags'].str.contains('classic eggs benedict')]
eggs_benedict_sentences.shape


eggs_benedict_sentences.head()

eggs_benedict_sentences.to_csv('../data/interim/eggs_benedict_sentences.csv', index = False)
