==> code/05.2-rating-features.ipynb <==
## Rating Features

### Questions:
1. Based on sentiment score, how are the key restaurant features scored at Mon Ami Gabi?  
    1. Food
    1. Service
    1. Cleanliness
    1. Ambiance
    1. Authenticity
    1. Portion Size

### Import libraries

import csv
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import pickle
import itertools as it
from nltk.sentiment.vader import SentimentIntensityAnalyzer
from IPython.display import clear_output
from time import time 

# For loading secret environment variables, e.g. postgres username and password
import os
from dotenv import load_dotenv, find_dotenv

# For reading from Postgres
from sqlalchemy import create_engine

# Display full content of column
pd.set_option('display.max_colwidth', 100)

### Set filepaths

# Directories
external_data_directory = os.path.join('..', 'data', 'external')
interim_data_directory = os.path.join('..', 'data', 'interim')
raw_data_directory = os.path.join('..', 'data', 'raw')

# Filepaths
review_csv_filepath = os.path.join(raw_data_directory, 
                                   'yelp_academic_dataset_review.csv')
restaurant_review_csv_filepath = os.path.join(interim_data_directory,
                                              'restaurant_review.csv')

mon_ami_gabi_filepath = os.path.join(interim_data_directory, 'mon_ami_gabi.csv')

### Set environment variables

public_ip = 'localhost'
username = 'postgres'
password = 'password'
port = '5432'
database = 'yelp'

# Construct database URL from environment variables
uri = f'postgresql://{username}:{password}@{public_ip}:{port}/{database}'

# Connection to Postgres database
engine = create_engine(uri)

### Get reviews for one restaurant by matching `business_id`

mon_ami_gabi_id = '4JNXUYY8wbaaDmk3BPzlWw'

%%time

mon_ami_gabi_id = '4JNXUYY8wbaaDmk3BPzlWw' 

SQL = f'''
SELECT date, stars, text, review_id, business_id, business_name
FROM reviews
WHERE business_id = '{mon_ami_gabi_id}'
'''

mon_ami_gabi = pd.read_sql(SQL, con = engine)

mon_ami_gabi.head()

mon_ami_gabi.shape

### Load Key Features

features = pickle.load(open('../data/interim/key_features.pk', 'rb'))
features.shape

features

### Preprocess restaurant reviews

#### Lowercase, drop punctuations

import re

def multiple_replace(string, replace_dict):
    '''
    Performs multiple string replacements at once
    '''
    pattern = re.compile('|'.join([re.escape(k) for k in sorted(replace_dict, key = len, reverse = True)]), flags = re.DOTALL)
    return pattern.sub(lambda x : replace_dict[x.group(0)], string)                  

stop_chars = []
replace_dict = {','   : '', 
                ';'   : '',
                '('   : '', 
                ')'   : '', 
                '\n' : ' ', 
                '!'   : '.',
                '?'   : '.'}

mon_ami_gabi['text'] = mon_ami_gabi['text'].apply(lambda text: multiple_replace(text.lower(), replace_dict))


mon_ami_gabi[['text']].head()

### For each item, extract sentence fragments

def find_term(word_list, term):
    '''
    Arguments:
    word_list : List of strings
    term      : List or string of words to search for
    
    Return:
    results : List of tuples of `start` and `end` indices of term in word_list,
              where `start` is the index of the first character in `term` in word_list,
              and `end` is the index of the last character in `term` in word_list.
    '''    
    # Check if word_list is a string or list
    if type(word_list) is str:
        word_list = word_list.split()
    elif type(word_list) is not list:
        print('Error: word_list must be a list or string.')
        return None

    # Check if term is a string or list    
    if type(term) is str:
        term = term.split()
    elif type(term) is not list:
        print('Error: term must be a list or string.')
        return None

    results = []
    term_length = len(term)

    # Find indices of term[0] in sentence
    for ind in (i for i, word in enumerate(word_list) if word == term[0]):
        # Check if rest of the term matches
        if word_list[ind:ind + term_length] == term:
            results.append((ind, ind+term_length-1))

    return results

def get_chunks(word_list, term, n_before = 5, n_after = 5):
    '''
    Arguments:
    word_list : List or string of words
    term      : List or string of words to search for
    before    : Number of characters to span before term
    after     : Number of characters to span after term   
    
    Builds a list of chunks containing term
    
    Return:
    chunks : List of chunks
    
    '''
    # Check if word_list is a string or list
    if type(word_list) is str:
        word_list = word_list.split()
    elif type(word_list) is not list:
        print('Error: word_list must be a list or string.')
        return None
    
    # Check if term is a string or list    
    if type(term) is str:
        term = term.split()
    elif type(term) is not list:
        print('Error: term must be a list or string.')
        return None    
    
    indices = find_term(word_list, term)
#     print(indices)
    chunks = []

    for start, end in indices:
        before = n_before
        after = n_after
        
        # Check if start index is near the beginning of the word_list
        if start < n_before:
            before = start
        # Check if end index is near the end of the word_list
        if end > len(word_list) - n_after:
            after = len(word_list) - end
            
        chunks.append(' '.join(word_list[start-before : end+after+1]))
        
    return chunks



### Extract Sentence Fragments for each food item

review_text = mon_ami_gabi[mon_ami_gabi['text'].str.contains('service')]

pd.set_option('display.max_colwidth', 200)
review_text[['text']].head()

%%time

# Instantiate features chunks list
features_chunks = {feature_id : [] for feature_id in features['id'] }

nrows = len(mon_ami_gabi)
# each review
for i, review in mon_ami_gabi.iterrows():
    if (i+1)%100==0:
        clear_output(wait = True)
        print(f'Parsing {i+1}/{nrows} reviews')
    
    # each sentence
    for sentence in review['text'].split('.'):
        # each feature item
        for _, item in features.iterrows():            
            # each feature variation
            for variation in item['variations']:
                # get sentence fragment if sentence contains feature item
                if variation in sentence:
                    chunks = get_chunks(sentence, variation, 7, 7)
                    features_chunks[item['id']].append(chunks)
                    break


#### Drop any empty lists

for k,v in features_chunks.items():
    features_chunks[k] = [c for c in v if c]

def flatten(superlist): 
    '''
    Arguments: 
    superlist : A list of list of strings.

    Requirements: 
    Each element in superlist must be a list.
    
    Return:
    A flattened list of strings.

    ex: 
    flatten([['a'], ['b', 'c'], ['d', 'e', 'f']])
    >> ['a', 'b', 'c', 'd', 'e', 'f']
    '''    
    return [item \
            for sublist in superlist \
            for item in sublist]

#### Flatten list of lists of feature chunks

# Flatten each food's list of chunks
features_chunks2 = {feature_id : flatten(chunks) for feature_id, chunks in features_chunks.items()}

# save for later
pickle.dump(features_chunks2, open("../data/interim/features_chunks.pk", "wb"))

### Predict sentiment score using `SentimentIntensityAnalyzer`

`SentimentIntensityAnalyzer` returns sentiment scores between `[0, 1]` for positive, negative, and neutral sentiments, `[-1, 1]` for the compound score.

def get_sentiments(docs):
    '''
    Returns a Dataframe of sentiment scores with columns:
    'compound', 'pos', 'neu', 'neg'
    
    For each doc, 'pos', 'neu', 'neg' scores add to 1.
    'compound' is an overall sentiment score of the doc.
    '''
    # Instantiate SentimentIntensityAnalyzer
    sia = SentimentIntensityAnalyzer()

    sentiments = []

    # Generate sentiment score for each review
    for doc in docs:
        sentiment = sia.polarity_scores(doc)
        sentiments.append(sentiment)

    return pd.DataFrame(sentiments)

service_chunks = flatten(features_chunks['service'])

service_sentiments = get_sentiments(service_chunks)

service_sentiments.head()

### Predict subjectivity score

A sentence can be **objective** (has no opinion: "I ordered the onion soup"), or **subjective** (has opinion: "The onion soup was great"). I want to predict the ratings of key features using only the subjective reviews. 

# TODO

### Predictions without dropping `compound=0`

features_sentiments = {}

for feature_id, chunks in features_chunks2.items():
    # Run sentiment analyzer if item has chunks, otherwise default to 0
    if len(chunks) > 0:
        features_sentiments[feature_id] = get_sentiments(chunks)['compound'].mean()
    else:
        features_sentiments[feature_id] = np.nan

### Convert Sentiment Score to Stars

Transform continuous `[-1.0, 1.0]` range to discrete `[1, 5]`  
Metric: RMSE of `stars` and `mean_compound`.

def transform_to_stars(n, lower = -1, upper = 1):
    '''
    Transform n to a discrete value between 1 to 5, inclusive.
    lower and upper set the bounds of n.
    '''
    if n < lower or n > upper:
        print(f'OutOfRangeError: Set n between [{lower},{upper}]')
        return None
    
    interval = np.round((upper - lower) / 9, 2)
    bins = [np.round(lower + i*interval, 2) for i in range(0,10)]
#     print(interval)
#     print(bins)
    
    # one star is the lowest rating, can't get 0.5 star
    n_categories = 9
    
    if n >= 0.5:
        return 5.0
    elif n >= 0.46:
        return 4.5
    elif n >= .42: 
        return 4.0
    elif n >= .38:
        return 3.5
    elif n >= .32:
        return 3.0
    elif n >= .28:        
        return 2.5
    elif n >= .24:
        return 2.0
    elif n >= .20:
        return 1.5
    else: 
        return 1.0
    
    for i in range(2,11):
        if bins[i-2] <= n <= bins[i-1]:
#             print(i)
            return i/2

    if n >= bins[-1]:
        return 5.0


def rescale_to_stars(scores, lower = -1, upper = 1):
    '''
    Rescale list of values to discrete values between 1 to 5, inclusive.

    Distributes the 5-star system to a uniform distribution of the input range.
    '''    
    interval = np.round((upper - lower) / 5, 2)
    bins = [np.round(lower + i*interval, 2) for i in range(0, 6)]

    return list(pd.cut(scores, bins = bins, labels = np.arange(1, 6, 1), include_lowest = True))


### Determine the overall sentiment score of each key feature

sentiment_df = pd.DataFrame(list(zip(features_sentiments.keys(), features_sentiments.values())), columns = ['feature_id', 'sentiment_score'])
sentiment_df

sentiment_df['stars_pred'] = sentiment_df['sentiment_score'].apply(lambda score : transform_to_stars(score))
sentiment_df['mentions'] = sentiment_df['feature_id'].apply(lambda feature_id : len(features_chunks[feature_id]))

## Overall sentiment of each key feature

sentiment_df.head()

## Food

food_df = pd.concat([pd.DataFrame(features_chunks2['food'], columns=['text']), get_sentiments(features_chunks2['food'])['compound']], axis = 1)
food_df.head()

### Most positive reviews on food

food_df.nlargest(5, 'compound')

### Most negative reviews on food 

food_df.nsmallest(5, 'compound')

## Service

service_df = pd.concat([pd.DataFrame(features_chunks2['service'], columns=['text']), get_sentiments(features_chunks2['service'])['compound']], axis = 1)
service_df.head()

### Most positive reviews on service 

service_df.nlargest(5, 'compound')

### Most negative reviews on service 

service_df.nsmallest(5, 'compound')

## Cleanliness

cleanliness_df = pd.concat([pd.DataFrame(features_chunks2['cleanliness'], columns=['text']), get_sentiments(features_chunks2['cleanliness'])['compound']], axis = 1)
cleanliness_df.head()

### Most positive reviews on cleanliness 

cleanliness_df.nlargest(5, 'compound')

### Most negative reviews on cleanliness 

cleanliness_df.nsmallest(5, 'compound')

## Ambiance

ambiance_df = pd.concat([pd.DataFrame(features_chunks2['ambiance'], columns=['text']), get_sentiments(features_chunks2['ambiance'])['compound']], axis = 1)
ambiance_df.head()

### Most positive reviews on ambiance

ambiance_df.nlargest(5, 'compound')

### Most negative reviews on ambiance

ambiance_df.nsmallest(5, 'compound')

## Authenticity

authenticity_df = pd.concat([pd.DataFrame(features_chunks2['authenticity'], columns=['text']), get_sentiments(features_chunks2['authenticity'])['compound']], axis = 1)
authenticity_df.head()

### Most positive reviews on authenticity

authenticity_df.nlargest(5, 'compound')

### Most negative reviews on authenticity

authenticity_df.nsmallest(5, 'compound')

Certain features are likely subject to bias. Reviews are more likely to write negatively about service and cleanliness.


sentiment_df.to_csv('../data/predictions/features_sentiment.csv', index=False)

### Predictions after dropping `compound=0`
I am treating chunks with `compound=0` as reviews with no opinion.

features_sentiments2 = {}

for feature_id, chunks in features_chunks2.items():
    # Run sentiment analyzer if item has chunks, otherwise default to 0
    if len(chunks) > 0:
        sentiment_score = get_sentiments(chunks)['compound']
        features_sentiments2[feature_id] = sentiment_score[sentiment_score != 0].mean()
    else:
        features_sentiments2[feature_id] = np.nan

sentiment_df2 = pd.DataFrame(list(zip(features_sentiments2.keys(), features_sentiments2.values())), columns = ['feature_id', 'sentiment_score'])


sentiment_df2['stars_pred'] = sentiment_df2['sentiment_score'].apply(lambda score : transform_to_stars(score))
sentiment_df2['mentions'] = sentiment_df2['feature_id'].apply(lambda feature_id : len(features_chunks[feature_id]))

sentiment_df2.to_csv('../data/predictions/features_sentiment_subjective.csv', index=False)

sentiment_df2
